% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ed.seqtests.R
\name{ed.seqtests}
\alias{ed.seqtests}
\title{Sequential Entropy Difference Tests for the GEVr Model}
\usage{
ed.seqtests(data, theta = NULL)
}
\arguments{
\item{data}{Data should be contain n rows, each a GEVr observation.}

\item{theta}{Estimate for theta in the vector form (loc, scale, shape). If NULL, uses the MLE from the largest (r-1) order statistics.}
}
\value{
A matrix containing the p-value and test statistics of the sequential tests.
}
\description{
Sequentially performs the entropy difference (ED) test for the dataset.
}
\details{
GEVr data (in matrix x) should be of the form x[i,1] > x[i, 2] > ... > x[i, r] for each observation i=1, ..., n.
}
\examples{
dat <- rgevr(500, 5, loc=0.5, scale=1, shape=0.5)
ed.seqtests(dat)
}

