% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ed.test.R
\name{ed.test}
\alias{ed.test}
\title{GEVr Entropy Difference Test}
\usage{
ed.test(data, theta = NULL)
}
\arguments{
\item{data}{Data should be contain n rows, each a GEVr observation.}

\item{theta}{Estimate for theta in the vector form (loc, scale, shape). If NULL, uses the MLE from the largest (r-1) order statistics.}
}
\value{
statistic Test statistic.

p.value P-value for the test.

theta Value of theta used in the test.
}
\description{
Goodness-of-fit test for GEVr using the difference in likelihood between GEVr and GEV(r-1). This can be used sequentially to test for the choice of r.
}
\details{
GEVr data (in matrix x) should be of the form x[i,1] > x[i, 2] > ... > x[i, r] for each observation i=1, ..., n.
}
\examples{
## This will test if the GEV2 distribution fits the data.
dat <- rgevr(100, 2, loc=0.5, scale=1, shape=0.5)

## Use the MLE from the GEV1 data by leaving theta input NULL.
result <- ed.test(dat)
}

