% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fortmax.R
\docType{data}
\name{fortmax}
\alias{fortmax}
\title{Top Ten Annual Precipitation: Fort Collins, Colorado}
\format{A data frame with 100 observations. Each year is considered an observation, with the top ten annual precipitation events.}
\source{
Colorado Climate Center, Colorado State University. This is the original data source containing daily precipitation data.
}
\usage{
data(fortmax)
}
\description{
Top ten annual precipitation events (inches) for one rain gauge
in Fort Collins, Colorado from 1900 through 1999. See
Katz et al. (2002) Sec. 2.3.1 for more information and analyses.
}
\examples{
data(fortmax)
y <- fortmax[, -1]
ed.seqtests(y)
}
\references{
Katz, R. W., Parlange, M. B. and Naveau, P. (2002) Statistics of extremes in hydrology. Advances in Water Resources, 25, 1287-1304.
}
\keyword{datasets}

