% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gev.multscore.R
\name{gev.multscore}
\alias{gev.multscore}
\title{GEVr Multiplier Score Test}
\usage{
gev.multscore(data, B, theta = NULL, information = c("expected",
  "observed"))
}
\arguments{
\item{data}{Data should be contain n rows, each a GEVr observation.}

\item{B}{Number of bootstrap replicates.}

\item{theta}{Estimate for theta in the vector form (loc, scale, shape). If NULL, an initial estimate is provided using PWM.}

\item{information}{To use observed or expected (default) information in the test.}
}
\value{
statistic Test statistic.

p.value P-value for the test.

theta Value of theta used in the test.
}
\description{
Fast weighted bootstrap alternative to the parametric bootstrap procedure for the GEVr score test.
}
\details{
GEVr data (in matrix x) should be of the form x[i,1] > x[i, 2] > ... > x[i, r] for each observation i=1, ..., n.
}
\examples{
data <- rgevr(500, 5, loc=0.5, scale=1, shape=0.3)
result <- gev.multscore(data, 1000)
}

