% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gev.pbscore.R
\name{gev.pbscore}
\alias{gev.pbscore}
\title{GEVr Parametric Bootstrap Score Test}
\usage{
gev.pbscore(data, B, information = c("expected", "observed"),
  allowParallel = FALSE, numCores = 1)
}
\arguments{
\item{data}{Data should be contain n rows, each a GEVr observation.}

\item{B}{Number of bootstrap replicates.}

\item{information}{To use observed or expected (default) information in the test.}

\item{allowParallel}{Should the bootstrap procedure be run in parallel or not. Defaults to false.}

\item{numCores}{If allowParallel is true, specify the number of cores to use.}
}
\value{
statistic Test statistic.

p.value P-value for the test.

theta Initial value of theta used in the test.
}
\description{
Parametric bootstrap score test procedure to assess goodness-of-fit to the GEVr distribution.
}
\details{
GEVr data (in matrix x) should be of the form x[i,1] > x[i, 2] > ... > x[i, r] for each observation i=1, ..., n.
}
\examples{
## Generate some data from GEVr
dat <- rgevr(200, 5, loc=0.5, scale=1, shape=0.5)
gev.pbscore(dat, 99)
}

