% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gev.spatpbscore.R
\name{gev.spatpbscore}
\alias{gev.spatpbscore}
\title{GEV1 (Spatially) Correlated Semi-Parametric Bootstrap Global Score Test}
\usage{
gev.spatpbscore(data, B, information = c("expected", "observed"),
  allowParallel = FALSE, numCores = 1)
}
\arguments{
\item{data}{Data should contain m columns, each data from a marginal GEV1 distribution. The number of rows
corresponds to the number of observations.}

\item{B}{Number of bootstrap replicates.}

\item{information}{To use observed or expected (default) information in the test.}

\item{allowParallel}{Should the bootstrap procedure be run in parallel or not. Defaults to false.}

\item{numCores}{If allowParallel is true, specify the number of cores to use.}
}
\value{
statistic Test statistic.

p.value P-value for the test.

theta Parameter estimates at each site.
}
\description{
Semi-parametric bootstrap score test procedure to assess global goodness-of-fit to the GEV1 distribution in
the presence of (spatial) correlation. Each marginal distribution (column) is assumed to be from the GEV1
distribution. The test statistic is a sum of marginal parametric bootstrap score test statistics (as in the
function gev.pbscore). In the prescence of correlation among marginal distributions, a semi-parametric
bootstrap procedure is done to preserve the dependence structure. See Heffernan and Tawn (2004) for more
details on that procedure.
}
\examples{
## Not run
## Generate data with max-stable dependence and GEV1 margins
## install.packages("SpatialExtremes")
## library(SpatialExtremes)
## n.site <- 20
## n.obs <- 50
## coord <- matrix(runif(2*n.site, 0, 10), ncol = 2)
## colnames(coord) <- c("lon", "lat")
## data <- rmaxstab(n.obs, coord, "gauss", cov11 = 100, cov12 = 25, cov22 = 220)
## param.loc <- -10 + 2 * coord[,2]
## param.scale <- 5 + 2 * coord[,1] + coord[,2]^2
## param.shape <- rep(0.2, n.site)
## for (i in 1:n.site) data[,i] <- frech2gev(data[,i], param.loc[i], param.scale[i], param.shape[i])
## gev.spatpbscore(data, 1000)
}
\references{
Heffernan, J. E., & Tawn, J. A. (2004). A conditional approach for multivariate extreme values (with discussion).
Journal of the Royal Statistical Society: Series B (Statistical Methodology), 66(3), 497-546.
}

