% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gevfit.R
\name{gevfit}
\alias{gevfit}
\title{Fits generalized extreme value distribution (GEV) to block maxima data.}
\usage{
gevfit(data, method = c("mle", "mps", "pwm"))
}
\arguments{
\item{data}{Data should be a numeric vector from the GEV distribution.}

\item{method}{Method of estimation - maximum likelihood (mle), probability weighted moments (pwm), and maximum product spacings (mps). Uses mle by default.}
}
\value{
A list describing the fit, including parameter estimates and standard errors for the mle and mps methods.
}
\description{
Fits generalized extreme value distribution (GEV) to block maxima data.
}
\examples{
data <- rgevr(500, 1,loc=0.5, scale=1, shape=0.3)
result <- gevfit(data, "mps")
}

