% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gevr.R
\name{dgevr}
\alias{dgevr}
\alias{rgevr}
\title{The GEVr Distribution}
\usage{
dgevr(x, loc = 0, scale = 1, shape = 0, log.d = FALSE)

rgevr(n, r, loc = 0, scale = 1, shape = 0)
}
\arguments{
\item{x}{Vector or matrix of observations. If x is a matrix, each row is taken to be a new observation.}

\item{loc}{Location parameter}

\item{scale}{Scale parameter}

\item{shape}{Shape Parameter}

\item{log.d}{Logical: Whether or not to return the log density (FALSE by default)}

\item{n}{Number of observations}

\item{r}{Number of order statistics for each observation}
}
\description{
Random number generation (rgevr) and density (dgevr) functions for the GEVr distribution with parameters loc, scale, and shape.
}
\details{
GEVr data (in matrix x) should be of the form x[i,1] > x[i, 2] > ... > x[i, r] for each observation i=1, ..., n.
}

