% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gevr.returnlevel.R
\name{gevr.returnlevel}
\alias{gevr.returnlevel}
\title{GEVr Return Level Estimate and Confidence Interval}
\usage{
gevr.returnlevel(data, period, conf = 0.95, method = c("delta", "profile"),
  opt = c("SANN", "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "Brent"))
}
\arguments{
\item{data}{Data should be a numeric vector (GEV1) or matrix from the GEVr distribution.}

\item{period}{The number of periods to use for the return level.}

\item{conf}{Confidence level. Defaults to 95 percent.}

\item{method}{The method to compute the confidence interval - either delta method (default) or profile likelihood.}

\item{opt}{Optimization method to maximize the profile likelihood if that is selected. The default (SANN) is slow, but works well.}
}
\value{
Estimate Estimated m-period return level.

CI Confidence interval for the m-period return level.

Period The period length used.
}
\description{
Computes m-period return level estimates, using either the delta method or profile likelihood. Can be for block maxima (GEV1) or GEVr.
}
\details{
Caution: The profile likelihood optimization may be slow (on the order of minutes).
}
\examples{
data <- rgevr(100, 5, loc=0.5, scale=1, shape=0.3)
## Compute 50-period return level.
gevr.returnlevel(data, 50)
}
\references{
http://www.mas.ncl.ac.uk/~nlf8/teaching/mas8391/background/chapter2.pdf
}

