% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gevscore.seqtests.R
\name{gevscore.seqtests}
\alias{gevscore.seqtests}
\title{Sequential Score Tests for the GEVr Model}
\usage{
gevscore.seqtests(data, nsim, method = c("pb", "mult"),
  information = c("expected", "observed"), allowParallel = FALSE,
  numCores = 1)
}
\arguments{
\item{data}{Data should be contain n rows, each a GEVr observation.}

\item{nsim}{Number of bootstrap simulations.}

\item{method}{Which test to run, multiplier or parametric (default) bootstrap.}

\item{information}{To use observed or expected (default) information in the test.}

\item{allowParallel}{Should the parametric boostrap procedure be run in parallel or not. Defaults to false.}

\item{numCores}{If allowParallel is true, specify the number of cores to use.}

\item{theta}{Estimate for theta in the vector form (loc, scale, shape). If NULL, uses the MLE.}
}
\value{
A matrix containing the test statistics and p-value results of the sequential tests.
}
\description{
Sequentially performs either the multiplier or parametric bootstrap score test for the GEVr model.
}
\details{
GEVr data (in matrix x) should be of the form x[i,1] > x[i, 2] > ... > x[i, r] for each observation i=1, ..., n.
}
\examples{
data <- rgevr(200, 3, loc=0.5, scale=1, shape=0.5)
gevscore.seqtests(data, 99)
}

