% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gpd.cvm.R
\name{gpd.cvm}
\alias{gpd.cvm}
\title{Generalized Pareto Distribution Cramer-von Mises Test}
\usage{
gpd.cvm(data, B, allowParallel = FALSE, numCores = 1)
}
\arguments{
\item{data}{Data should be in vector form, assumed to be from the GP distribution.}

\item{B}{Number of bootstrap replicates.}

\item{allowParallel}{Should the bootstrap procedure be run in parallel or not. Defaults to false.}

\item{numCores}{If allowParallel is true, specify the number of cores to use.}
}
\value{
statistic Test statistic.

p.value P-value for the test.

theta Estimated value of theta for the initial data.
}
\description{
Cramer-von Mises goodness-of-fit test for the Generalized Pareto distribution. Critical values are generated via parametric bootstrap.
}
\examples{
## Generate some data from GPD
dat <- rgpd(200, 0, 1, 0.2)
gpd.cvm(dat, 100)
}
\references{
Choulakian, V., & Stephens, M. A. (2001). Goodness-of-fit tests for the generalized Pareto distribution. Technometrics, 43(4), 478-484.
}

