% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gpd.impb.R
\name{gpd.impb}
\alias{gpd.impb}
\title{GPD Bootstrapped Information Matrix Test}
\usage{
gpd.impb(data, inner, outer, allowParallel = FALSE, numCores = 1)
}
\arguments{
\item{data}{Data should be in vector form.}

\item{inner}{Number of bootstrap replicates for the covariance estimate.}

\item{outer}{Number of bootstrap replicates for critical values.}

\item{allowParallel}{Should the outer bootstrap procedure be run in parallel or not. Defaults to false.}

\item{numCores}{If allowParallel is true, specify the number of cores to use.}
}
\value{
statistic Test statistic.

p.value P-value for the test.

theta Estimate of theta for the initial dataset.

effective_bootnum Effective number of outer bootstrap replicates used (if some did not converge).
}
\description{
Runs the IM Test using a two-step iterative procedure, to boostrap the covariance estimate and critical values. (See reference for details.)
}
\examples{
## Generate some data from GPD
dat <- rgpd(200, 0, 1, -0.2)
gpd.impb(dat, 50, 99)
}
\references{
Dhaene, G., & Hoorelbeke, D. (2004). The information matrix test with bootstrap-based covariance matrix estimation. Economics Letters, 82(3), 341-347.
}

