% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gpd.multscore.R
\name{gpd.multscore}
\alias{gpd.multscore}
\title{GPD Multiplier Score Test}
\usage{
gpd.multscore(data, B, theta = NULL, information = c("expected",
  "observed"))
}
\arguments{
\item{data}{Data should be in vector form.}

\item{B}{Number of bootstrap replicates.}

\item{theta}{Estimate for theta in the vector form (scale, shape). If NULL, uses probability weighted moments as an initial estimate.}

\item{information}{To use observed or expected (default) information in the test.}
}
\value{
statistic Test statistic.

p.value P-value for the test.

theta Value of theta used in the test.
}
\description{
Fast weighted bootstrap alternative to the parametric bootstrap procedure for the Generalized Pareto score test.
}
\examples{
data <- rgpd(100, 0, 1, 0.25)
gpd.multscore(data, 1000)
}

