% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gpdfit.R
\name{gpdfit}
\alias{gpdfit}
\title{Fits the generalized pareto distribution to data}
\usage{
gpdfit(data, threshold = NA, nextremes = NA, method = c("mle", "mps",
  "pwm"), information = c("observed", "expected"))
}
\arguments{
\item{data}{Data should be a numeric vector from the GPD distribution.}

\item{threshold}{A threshold value (either this or the number of extremes must be given, but not both).}

\item{nextremes}{Number of upper extremes to be used.}

\item{method}{Method of estimation - maximum likelihood (mle), probability weighted moments (pwm), and maximum product spacings (mps). Uses mle by default.}

\item{information}{Whether standard errors should be calculated via observed or expected information. For probability weighted moments, only expected information will be used if possible.}
}
\value{
A list describing the fit, including parameter estimates and standard errors.
}
\description{
The base code for this function is taken from the R package evir. See citation below. The addition here includes estimation via Maximum Product Spacings (MPS).
}
\references{
Pfaff, Bernhard, Alexander McNeil, and A. Stephenson. "evir: Extreme Values in R." R package version (2012): 1-7.
}

