% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gpdthresh.seqtests.R
\name{gpdthresh.seqtests}
\alias{gpdthresh.seqtests}
\title{GPD Multiple Threshold GoF Testing}
\usage{
gpdthresh.seqtests(data, thresholds = NA, nextremes = NA,
  method = c("cvm", "ad", "pbscore", "multscore", "imasym", "impb"),
  nsim = NULL, inner = NULL, outer = NULL, information = c("expected",
  "observed"), allowParallel = FALSE, numCores = 1)
}
\arguments{
\item{data}{Orginal, full data in vector form.}

\item{thresholds}{A set of threshold values (either this or a set of the number of extremes must be given, but not both).}

\item{nextremes}{A set of the number of upper extremes to be used.}

\item{method}{Which test to run to sequentially test the thresholds. Must be one of 'cvm', 'ad', pbscore', 'multscore', 'imasym', or 'impb'.}

\item{nsim}{Number of boostrap replicates for the 'cvm', 'ad', 'pbscore', 'multscore', and 'imasym' tests.}

\item{inner}{Number of inner boostrap replicates if 'impb' test is chosen.}

\item{outer}{Number of outer boostrap replicates if 'impb' test is chosen.}

\item{information}{To use observed or expected (default) information for the 'pbscore' and 'multscore' tests.}

\item{allowParallel}{If selected, should the 'cvm', 'ad', 'pbscore', or 'impb' procedure be run in parallel or not. Defaults to false.}

\item{numCores}{If allowParallel is true, specify the number of cores to use.}
}
\value{
A matrix containing the thresholds used, the number of observations above each threshold, and the corresponding test statistics and p-values.
}
\description{
Wrapper functions to test multiple thresholds for goodness-of-fit to the Generalized Pareto model. Can choose which test to run.
}

